tetragon/windows: Port reader/proc package on Windows
This PR ports multiple utility functions related to querying
process's user, group, login ids into Windows using Win32 APIs
